test_picking_algorithms/largest_gap.py
def largest_gap_picking_tecnnique(sku_objects):
    """
    Implementation of the Largest Gap order picking algorithm.

    Parameters:
        sku_objects (list of SKU): Each SKU object has a pickup_location attribute that represents the 2D coordinates (x, y) of a picking location.

    Returns:
        list of SKU: The sorted SKU objects based on the Largest Gap algorithm.
    """
    if len(sku_objects) <= 1:
        return sku_objects

    # Function to calculate the distance between two points (Euclidean distance)
    def distance(point1, point2):
        return ((point1[0] - point2[0]) ** 2 + (point1[1] - point2[1]) ** 2) ** 0.5

    # Sort the SKU objects based on the x-coordinate of the pickup_location
    sku_objects.sort(key=lambda sku: sku)

    # Calculate the distances between consecutive points
    distances = [distance(sku_objects[i], sku_objects[i +
                          1]) for i in range(len(sku_objects) - 1)]

    # Find the index of the largest gap (maximum distance between two consecutive points)
    max_gap_index = distances.index(max(distances))

    # Reorder the SKU objects starting from the point after the largest gap
    sorted_sku_objects = sku_objects[max_gap_index +
                                     1:] + sku_objects[:max_gap_index + 1]

    return sorted_sku_objects
